# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"PageMetadata": {
				"properties": {
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"totalElements": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageSize",
					"totalElements",
					"totalPages",
					"currentPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Country": {
				"properties": {
					"alpha2Code": {
						"type": "string"
					},
					"alpha3Code": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"subregion": {
						"type": "string"
					},
					"commonName": {
						"type": "string"
					},
					"officialName": {
						"type": "string"
					}
				},
				"required": [
					"alpha2Code",
					"alpha3Code",
					"region",
					"subregion",
					"commonName",
					"officialName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"founder": {
						"type": "string"
					},
					"yearFounded": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"fullName",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PhotoSchema": {
				"allOf": [
					{
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					{
						"properties": {
							"CC": {
								"type": "number",
								"format": "double",
								"description": "The CC license version number (if not CC public domain)"
							},
							"CCType": {
								"type": "string",
								"enum": [
									"BY",
									"BY-SA",
									"BY-ND",
									"BY-NC",
									"BY-NC-ND",
									"Public Domain"
								],
								"description": "The CC license type."
							},
							"siteURL": {
								"type": "string",
								"description": "Web site where the photo is. Not to be confused with the URL of the photo itself. In this, in addition to the photo, there is usually other information such as the author, date of capture..."
							},
							"site": {
								"type": "string",
								"description": "Web site name of the URL provided."
							},
							"author": {
								"type": "string",
								"description": "Author name of the photo."
							},
							"description": {
								"type": "string",
								"description": "Description of the picture."
							},
							"url": {
								"type": "string",
								"description": "The URL of the photo."
							},
							"embedId": {
								"type": "number",
								"format": "double",
								"description": "ID of the image to embed via SmartFrame.io"
							}
						},
						"type": "object"
					}
				],
				"description": "A valid photo object, with license info."
			},
			"Driver": {
				"description": "Object that contains all the basic info about a driver",
				"properties": {
					"id": {
						"type": "string",
						"description": "The unique identifier."
					},
					"name": {
						"type": "string",
						"description": "The name."
					},
					"firstName": {
						"type": "string",
						"description": "The first name."
					},
					"lastName": {
						"type": "string",
						"description": "The last name."
					},
					"fullName": {
						"type": "string",
						"description": "The full name."
					},
					"abbreviation": {
						"type": "string",
						"description": "The abbreviation."
					},
					"permanentNumber": {
						"type": "string",
						"nullable": true,
						"description": "The permanent number."
					},
					"gender": {
						"type": "string",
						"enum": [
							"MALE",
							"FEMALE"
						],
						"description": "The gender."
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time",
						"description": "The date of birth."
					},
					"dateOfDeath": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "The date of death."
					},
					"placeOfBirth": {
						"type": "string",
						"description": "The place of birth."
					},
					"countryOfBirth": {
						"$ref": "#/components/schemas/Country"
					},
					"nationalityCountry": {
						"$ref": "#/components/schemas/Country"
					},
					"secondNationalityCountry": {
						"$ref": "#/components/schemas/Country"
					},
					"photo": {
						"$ref": "#/components/schemas/PhotoSchema"
					}
				},
				"required": [
					"id",
					"name",
					"firstName",
					"lastName",
					"fullName",
					"abbreviation",
					"permanentNumber",
					"gender",
					"dateOfBirth",
					"dateOfDeath",
					"placeOfBirth",
					"countryOfBirth",
					"nationalityCountry"
				],
				"type": "object",
				"additionalProperties": false
			},
			"qualyFormat": {
				"enum": [
					"FOUR_LAPS",
					"TWO_SESSION",
					"ONE_SESSION",
					"ONE_LAP",
					"AGGREGATE",
					"KNOCKOUT",
					"ELIMINATION",
					"SPRINT_RACE"
				],
				"type": "string"
			},
			"Circuit": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"previousNames": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"RACE",
							"ROAD",
							"STREET"
						]
					},
					"placeName": {
						"type": "string"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"latitude": {
						"type": "number",
						"format": "double",
						"description": "The latitude of the location"
					},
					"longitude": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "The longitude of the location"
					}
				},
				"required": [
					"id",
					"name",
					"fullName",
					"previousNames",
					"type",
					"placeName",
					"country",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GrandPrix": {
				"properties": {
					"id": {
						"type": "string",
						"description": "The unique identifier."
					},
					"name": {
						"type": "string",
						"description": "The name."
					},
					"fullName": {
						"type": "string"
					},
					"shortName": {
						"type": "string"
					},
					"countryId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"fullName",
					"shortName",
					"countryId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Event": {
				"properties": {
					"id": {
						"type": "string"
					},
					"raceDate": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"scheduledLaps": {
						"type": "number",
						"format": "double"
					},
					"qualyFormat": {
						"$ref": "#/components/schemas/qualyFormat"
					},
					"posterURL": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double",
						"description": "The year (or F1 session) of the event. Obtained from the ID in the constructor."
					},
					"round": {
						"type": "number",
						"format": "double",
						"description": "The round number of the event in its year. Obtained from the ID in the constructor."
					},
					"circuit": {
						"$ref": "#/components/schemas/Circuit",
						"description": "The full circuit object. Got in the contructor"
					},
					"grandPrix": {
						"$ref": "#/components/schemas/GrandPrix",
						"description": "The full grand prix object. Got in the contructor"
					}
				},
				"required": [
					"id",
					"raceDate",
					"name",
					"scheduledLaps",
					"qualyFormat",
					"posterURL",
					"year",
					"round",
					"circuit",
					"grandPrix"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SeasonEntrant": {
				"properties": {
					"id": {
						"type": "string"
					},
					"season": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					}
				},
				"required": [
					"id",
					"season",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TyreManufacturer": {
				"properties": {
					"id": {
						"type": "string",
						"description": "The unique identifier."
					},
					"name": {
						"type": "string",
						"description": "The name."
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"primaryColor": {
						"type": "string"
					},
					"secondaryColor": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"country",
					"primaryColor",
					"secondaryColor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimedSessionResult": {
				"description": "Common interface for timed sessions, like free practices, warm ups...",
				"properties": {
					"driver": {
						"$ref": "#/components/schemas/Driver",
						"description": "The full driver object. Got in the contructor"
					},
					"race": {
						"$ref": "#/components/schemas/Event",
						"description": "The full race event object. Got in the contructor"
					},
					"seasonEntrant": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SeasonEntrant"
							}
						],
						"nullable": true
					},
					"entrantName": {
						"type": "string"
					},
					"chassisManufacturer": {
						"$ref": "#/components/schemas/Company"
					},
					"chassisName": {
						"type": "string"
					},
					"engineManufacturer": {
						"$ref": "#/components/schemas/Company"
					},
					"engineName": {
						"type": "string"
					},
					"tyreManufacturer": {
						"$ref": "#/components/schemas/TyreManufacturer"
					},
					"driverNumber": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"positionOrder": {
						"type": "number",
						"format": "double"
					},
					"positionText": {
						"type": "string"
					},
					"position": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"laps": {
						"type": "number",
						"format": "double",
						"description": "Number of laps completed in this session."
					},
					"time": {
						"type": "number",
						"format": "double",
						"description": "Time that gives the driver his/her position at the end of this session."
					}
				},
				"required": [
					"driver",
					"race",
					"seasonEntrant",
					"entrantName",
					"chassisManufacturer",
					"chassisName",
					"engineManufacturer",
					"tyreManufacturer",
					"positionOrder",
					"positionText",
					"position",
					"time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderDir": {
				"type": "string",
				"enum": [
					"ASC",
					"DESC"
				]
			},
...
			"ErrorMessage_404_": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							404
						],
						"nullable": false
					},
					"code": {
						"type": "string"
					},
					"details": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"code",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend-f1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/free-practices/{session}/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain session of free practice",
				"tags": [
					"Free Practices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"fp1",
								"fp2",
								"fp3",
								"fp4",
								"warm-up"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/free-practices/{session}/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain session of free practice",
				"tags": [
					"Free Practices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"fp1",
								"fp2",
								"fp3",
								"fp4",
								"warm-up"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/pre-qualifyings/results": {
			"get": {
				"operationId": "GetPreQualifyingsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/TimedSessionResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Return driver results of a specific pre-qualifying session based on some filters",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId"
							]
						}
					}
				]
			}
		},
		"/pre-qualifyings/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain pre-qualifying",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pre-qualifyings/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a pre-qualifying session",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/qualifyings/{session}/results": {
			"get": {
				"operationId": "GetSessionsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/TimedSessionResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Return driver results of a specific qualifying session based on some filters",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId"
							]
						}
					}
				]
			}
		},
		"/qualifyings/{session}/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain session of qualifying",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/qualifyings/{session}/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain session of qualifying",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}